refactor(#211): extract WorldSummary structs to game_summary.h, trim game.h to 142 lines
Move WorldPeerSummary, WorldPlayerSummary, WorldSummary out of game.h
into a standalone game_summary.h. game.h shrinks from 174 → 142 lines.
Update controldispatch.cpp and game_loop.cpp to drop the Game:: prefix.
