+++ clients/silencer/src/net/controldispatch.cpp
@@ -143,7 +143,7 @@ static nlohmann::json InGameUiControlResultToJson(
 	return r;
 }
 
-static nlohmann::json WorldSummaryToJson(const Game::WorldSummary& summary){
+static nlohmann::json WorldSummaryToJson(const WorldSummary& summary){
 	nlohmann::json r;
 	r["map"] = summary.map;
 	r["peers"] = summary.peers;
@@ -156,7 +156,7 @@ static nlohmann::json WorldSummaryToJson(const Game::WorldSummary& summary){
 	r["spectator_freecam"] = summary.spectatorFreecam;
 
 	nlohmann::json peerlist = nlohmann::json::array();
-	for(const Game::WorldPeerSummary& peer : summary.peerList){
+	for(const WorldPeerSummary& peer : summary.peerList){
 		nlohmann::json p;
 		p["id"] = peer.id;
 		p["accountid"] = peer.accountId;
@@ -170,7 +170,7 @@ static nlohmann::json WorldSummaryToJson(const Game::WorldSummary& summary){
 	r["peerlist"] = peerlist;
 
 	nlohmann::json players = nlohmann::json::array();
-	for(const Game::WorldPlayerSummary& player : summary.players){
+	for(const WorldPlayerSummary& player : summary.players){
 		nlohmann::json p;
 		p["id"] = player.id;
 		p["hp"] = player.hp;
